+++ src/main/java/frc/robot/XboxOI.java
@@ -1,90 +1,193 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2018-2019 FIRST. All Rights Reserved.                        */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
 package frc.robot;
 
-import java.util.function.BooleanSupplier;
-
+import com.nerdherd.lib.drivetrain.auto.DriveDistanceMotionMagic;
+import com.nerdherd.lib.drivetrain.auto.DriveStraightContinuous;
+import com.nerdherd.lib.drivetrain.auto.ResetDriveEncoders;
+import com.nerdherd.lib.drivetrain.auto.ResetGyro;
+import com.nerdherd.lib.drivetrain.characterization.DriveCharacterizationTest;
 import com.nerdherd.lib.drivetrain.shifting.ShiftHigh;
 import com.nerdherd.lib.drivetrain.shifting.ShiftLow;
 import com.nerdherd.lib.motor.commands.ResetSingleMotorEncoder;
 import com.nerdherd.lib.motor.commands.SetMotorPower;
+import com.nerdherd.lib.motor.single.SingleMotorMechanism;
+import com.nerdherd.lib.oi.DefaultOI;
 import com.nerdherd.lib.oi.XboxDriverOI;
 
+import edu.wpi.first.wpilibj.Joystick;
+import edu.wpi.first.wpilibj.Sendable;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj.GenericHID.Hand;
 import edu.wpi.first.wpilibj.GenericHID.RumbleType;
+import edu.wpi.first.wpilibj.geometry.Pose2d;
+import edu.wpi.first.wpilibj.geometry.Rotation2d;
+
+// import org.graalvm.compiler.lir.aarch64.AArch64Move.StoreConstantOp;
+
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
-import edu.wpi.first.wpilibj2.command.button.Button;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
+import edu.wpi.first.wpilibj2.command.button.Button;
+import frc.robot.commands.InGameResetHood;
+import frc.robot.commands.auto.AutoLineIntoTrench;
+import frc.robot.commands.auto.BarrelRacing;
+import frc.robot.commands.auto.Bounce;
+import frc.robot.commands.auto.BounceBackwards;
+import frc.robot.commands.auto.GalacticPathABlue;
+import frc.robot.commands.auto.GalacticPathARed;
+import frc.robot.commands.auto.GalacticPathBBlue;
+import frc.robot.commands.auto.GalacticPathBRed;
+import frc.robot.commands.auto.GalacticReal;
+import frc.robot.commands.auto.Lightspeed;
+import frc.robot.commands.auto.PowerPort;
+import frc.robot.commands.auto.Slalom;
+import frc.robot.commands.auto.TestRamsete;
+import frc.robot.commands.auto.TestRamseteTurn;
+import frc.robot.commands.auto.Trapezoid;
+// import frc.robot.commands.AutolineShot;
+// import frc.robot.commands.ShootBall;
+// import frc.robot.commands.ShootBallTemp;
+// import frc.robot.commands.ShootBallTempStop;
+// import frc.robot.commands.TrenchShot;
+// import frc.robot.commands.WallShot;
+// import frc.robot.commands.auto.StealTwoEnemyTrench;
 import frc.robot.commands.climber.ClimberLift;
 import frc.robot.commands.climber.ClimberReady;
 import frc.robot.commands.intake.IntakeBalls;
 import frc.robot.commands.intake.Stow;
 import frc.robot.commands.other.SetAngle;
+import frc.robot.commands.other.ShootBallTemp;
 import frc.robot.commands.shooting.AutolineShot;
 import frc.robot.commands.shooting.RendezvousShot;
 import frc.robot.commands.shooting.ShootBall;
+// import frc.robot.commands.ShootBallTemp;
+// import frc.robot.commands.ShootBallTempStop;
 import frc.robot.commands.shooting.TrenchShot;
 import frc.robot.commands.shooting.WallShot;
+import frc.robot.commands.vision.DistanceToAngle;
+import frc.robot.commands.vision.TurnToAngleLime;
+import frc.robot.constants.VisionConstants;
+import frc.robot.subsystems.Indexer.IndexerState;
 
+/**
+ * Xbox OI Class
+ */
 public class XboxOI extends XboxDriverOI {
-    
-    private Button m_shiftLow, m_shiftHigh, m_intake, m_outtake, m_rightEncoder, m_leftEncoder, m_stow; // driver
-    private JoystickButton m_startShooting, m_climbReady, m_climbLift, m_wallShot, m_autolineShot, m_hoodAngle, m_rendezvousShot, m_trenchShot; // operator
-    
+
+    public JoystickButton startShooting_2, startShootingOld_3, trenchShot_7, autolineShot_9, stow_10, wallShot_11, autoDistance_12, hoodAngle_5, outtake_6, togglePipeline_4, rendezvousShot_8, outtakeBrushes_8, intake_1;
+
+    public JoystickButton shiftLow_A, shiftHigh_B, climbReady_X, climbLift_Y, resetEncoders_LB, ploughIntake_RB;
+    public Button intake_RT, turnToAngle_LT;
+
+    public final int BUTTON_A = 1, BUTTON_B = 2, BUTTON_X = 3, BUTTON_Y = 4, BUTTON_LB = 5, BUTTON_RB = 6, BUTTON_BACK = 7, BUTTON_START = 8, BUTTON_LEFT_STICK = 9, BUTTON_RIGHT_STICK = 10;
+
     private double m_triggerThreshold = 0.25;
 
     public XboxOI() {
-        this(0.1);
+        this(0);
     }
 
     public XboxOI(double deadband) {
-        m_shiftLow = new JoystickButton(super.driverController, XboxController.Button.kA.value);
-        m_shiftHigh = new JoystickButton(super.driverController, XboxController.Button.kB.value);
-        m_intake = new JoystickButton(super.driverController, XboxController.Button.kBumperRight.value);
-        m_outtake = new JoystickButton(super.driverController, XboxController.Button.kBumperLeft.value);
-        m_leftEncoder = new JoystickButton(super.driverController, XboxController.Button.kBack.value);
-        m_rightEncoder = new JoystickButton(super.driverController, XboxController.Button.kStart.value);
+        super(deadband);
+        intake_1 = new JoystickButton(super.operatorJoy, 1);
+        startShooting_2 = new JoystickButton(super.operatorJoy, 2);
+        startShootingOld_3 = new JoystickButton(super.operatorJoy, 3);
+        trenchShot_7 = new JoystickButton(super.operatorJoy, 7);
+        autolineShot_9 = new JoystickButton(super.operatorJoy, 9);
+        stow_10 = new JoystickButton(super.operatorJoy, 10);
+        wallShot_11 = new JoystickButton(super.operatorJoy, 11);
+        autoDistance_12 = new JoystickButton(super.operatorJoy, 12);
+        hoodAngle_5 = new JoystickButton(super.operatorJoy, 5);
+        outtake_6 = new JoystickButton(super.operatorJoy, 6);
+        togglePipeline_4 = new JoystickButton(super.operatorJoy, 4);
+        rendezvousShot_8 = new JoystickButton(super.operatorJoy, 8);
+        outtakeBrushes_8 = new JoystickButton(super.operatorJoy, 8);
 
-        // create an analog trigger for L and R
-        m_stow = new Button() {
+        shiftLow_A = new JoystickButton(super.driverController, BUTTON_A);
+        shiftHigh_B = new JoystickButton(super.driverController, BUTTON_B);
+        climbReady_X = new JoystickButton(super.driverController, BUTTON_X);
+        climbLift_Y = new JoystickButton(super.driverController, BUTTON_Y);
+        resetEncoders_LB = new JoystickButton(super.driverController, BUTTON_LB);
+        ploughIntake_RB = new JoystickButton(super.driverController, BUTTON_RB);
+        intake_RT = new Button() {
             @Override
             public boolean get() {
                 return getTrigger(Hand.kRight);
             }
         };
+        turnToAngle_LT = new Button() {
+            @Override
+            public boolean get() {
+                return getTrigger(Hand.kLeft);
+            }
+        };
 
-        m_shiftLow.whenPressed(new ShiftLow(Robot.drive));
-        m_shiftHigh.whenPressed(new ShiftHigh(Robot.drive));
-        m_intake.whenPressed(new IntakeBalls());
-        m_outtake.whenPressed(new SetMotorPower(Robot.intakeRoll, -0.75)
-            .alongWith(
+        intake_1.whenPressed(new IntakeBalls());
+        startShooting_2.whileHeld(new ShootBall());
+        startShootingOld_3.whileHeld(new ShootBallTemp());
+        togglePipeline_4.whenPressed(new InstantCommand(() -> Robot.limelight.togglePipeline()));
+        hoodAngle_5.whenPressed(new SetAngle());
+        outtake_6.whenPressed(new SetMotorPower(Robot.intakeRoll, -0.75).alongWith(
                 new InstantCommand(() -> Robot.hopper.setPowerWithoutTop(-0.4, -0.8)),
-                new SetMotorPower(Robot.index, -0.33), 
-                new InstantCommand(() -> Robot.hopper.setTopHopperPower(0.41))
-            ));
-        m_rightEncoder.whenPressed(Robot.hoodReset);
-        m_leftEncoder.whenPressed(Robot.hoodReset);
-        m_stow.whenPressed(new Stow());
+                new SetMotorPower(Robot.index, -0.33), new InstantCommand(() -> Robot.hopper.setTopHopperPower(0.41))));
+        trenchShot_7.whenPressed(new TrenchShot());
+        rendezvousShot_8.whenPressed(new RendezvousShot());
+        outtakeBrushes_8.whenHeld(new InstantCommand(() -> Robot.hopper.setTopHopperPower(-0.41)));
+        autolineShot_9.whenPressed(new AutolineShot());
+        stow_10.whenPressed(new Stow());
+        wallShot_11.whenPressed(new WallShot());
+        autoDistance_12.whenPressed(new DistanceToAngle());
 
-        m_startShooting = new JoystickButton(super.operatorJoy, 1);
-        m_hoodAngle = new JoystickButton(super.operatorJoy, 5);
-        m_climbReady = new JoystickButton(super.operatorJoy, 3);
-        m_climbLift = new JoystickButton(super.operatorJoy, 4);
-        m_wallShot = new JoystickButton(super.operatorJoy, 7);
-        m_autolineShot = new JoystickButton(super.operatorJoy, 8);
-        m_rendezvousShot = new JoystickButton(super.operatorJoy, 11);
-        m_trenchShot = new JoystickButton(super.operatorJoy, 12);
-
-        m_startShooting.whileHeld(new ShootBall());
-        m_hoodAngle.whenPressed(new SetAngle());
-        m_climbReady.whenPressed(new ClimberReady());
-        m_climbLift.whenPressed(new ClimberLift());
-        m_wallShot.whenPressed(new WallShot());
-        m_autolineShot.whenPressed(new AutolineShot());
-        m_rendezvousShot.whenPressed(new RendezvousShot());
-        m_trenchShot.whenPressed(new TrenchShot());
-
+        resetEncoders_LB.whenPressed(Robot.hoodReset);
+        ploughIntake_RB.whenPressed(new SetMotorPower(Robot.intakeRoll, -0.75));
+        shiftHigh_B.whenPressed(new ShiftHigh(Robot.drive));
+        shiftLow_A.whenPressed(new ShiftLow(Robot.drive));
+        turnToAngle_LT.whileHeld(new TurnToAngleLime(VisionConstants.kRotP_lime)); // .009 before
+        intake_RT.whenPressed(new IntakeBalls());
+        climbReady_X.whenPressed(new ClimberReady());
+        climbLift_Y.whileHeld(new ClimberLift());
+        
+        SmartDashboard.putData("Reset indexer",
+                new InstantCommand(() -> Robot.index.indexerState = IndexerState.EMPTY));
+        SmartDashboard.putData("cLIMBER UP", new ClimberReady());
+        SmartDashboard.putData("Climber Lift", new ClimberLift());
+        SmartDashboard.putData("Full Send", new DriveStraightContinuous(Robot.drive, 200000, 0.125));
+        SmartDashboard.putData("FJFJFJFJFJFJFJFJFJF", new GalacticReal(Robot.drive));
+        SmartDashboard.putData("PowerPort", new PowerPort(Robot.drive));
+        SmartDashboard.putData("REAL TRAP",
+                new InstantCommand(() -> Robot.drive.setPositionMotionMagic(200000, 200000, 2000, 1500)));
+        SmartDashboard.putData("trapezoid", new Trapezoid(10, Robot.drive));
+        SmartDashboard.putData("Slalom Drive", new Slalom(Robot.drive));
+        SmartDashboard.putData("Ramp Test", new DriveCharacterizationTest(Robot.drive, 0.25));
+        SmartDashboard.putData("Bounce Drive", new Bounce(Robot.drive));
+        SmartDashboard.putData("Bounce Backwards Drive", new BounceBackwards(Robot.drive));
+        SmartDashboard.putData("Lightspeed Drive", new Lightspeed(Robot.drive));
+        SmartDashboard.putData("Barrel Racing Drive", new BarrelRacing(Robot.drive));
+        SmartDashboard.putData("Galactic Path A Red Drive", new GalacticPathARed(Robot.drive));
+        SmartDashboard.putData("Galactic Path A Blue Drive", new GalacticPathABlue(Robot.drive));
+        SmartDashboard.putData("Galactic Path B Red Drive", new GalacticPathBRed(Robot.drive));
+        SmartDashboard.putData("Galactic Path B Blue Drive", new GalacticPathBBlue(Robot.drive));
+        SmartDashboard.putData("Test Ramsete", new TestRamsete(Robot.drive));
+        SmartDashboard.putNumber("Left Voltage", Robot.drive.getLeftOutputVoltage());
+        SmartDashboard.putData("Config Slalom Heading",
+                new InstantCommand(() -> Robot.drive.setPose(new Pose2d(0.762, -0.762, new Rotation2d(Math.PI / 2)))));
+        SmartDashboard.putData("Config Galactic Heading",
+                new InstantCommand(() -> Robot.drive.setPose(new Pose2d(2.286, -0.762, new Rotation2d(Math.PI / 2)))));
+        SmartDashboard.putData("4VShooter", new SetMotorPower(Robot.shooter, 0.33));
+        SmartDashboard.putData("SetHoodFF lime -> deg", new DistanceToAngle());
         SmartDashboard.putData("ResetHoodEncoder", new ResetSingleMotorEncoder(Robot.hood));
+        SmartDashboard.putData("Reset Encoders", new ResetDriveEncoders(Robot.drive));
+        SmartDashboard.putData("Reset XY", new InstantCommand(() -> Robot.drive.resetXY()));
+        SmartDashboard.putData("Reset Gyro", new ResetGyro(Robot.drive));
+        SmartDashboard.putData("Test Ramsete", new TestRamsete(Robot.drive));
+        SmartDashboard.putData("Test Ramsete Turn", new TestRamseteTurn(Robot.drive));
+        SmartDashboard.putData("InGameResetEncoders", new InGameResetHood());
     }
 
     public boolean getRawButton(int n) {
@@ -102,5 +205,4 @@ public class XboxOI extends XboxDriverOI {
     public boolean getTrigger(Hand hand) {
         return getTriggerAxis(hand) >= m_triggerThreshold;
     }
-
 }
